refact: API and authentication system
- Renamed project from "api" to "synthgenie" in pyproject.toml.
- Updated dependencies to newer versions in pyproject.toml.
- Added authentication routes for API key management including generation, revocation, and usage statistics.
- Created models and schemas for API key handling.
- Implemented services for API key creation, validation, and revocation.
- Removed old agent routes and integrated new synthesizer routes for Digitone and Sub37.
- Updated synthesizer tools to use shared schemas for responses.
- Adjusted FastAPI app to include new routers and middleware.
- Cleaned up unused files and organized the project structure.
