--- api/synthgenie/routes/__init__.py
@@ -1,1 +1,0 @@
-# Routes package initialization
