+++ api/synthgenie/synthesizers/digitone/tools/amp_fx_tool.py
@@ -1,6 +1,6 @@
 from pydantic_ai import RunContext
 
-from synthgenie.schemas.agent import SynthGenieResponse
+from synthgenie.synthesizers.shared.schemas.agent import SynthGenieResponse
 
 
 # Amp Functions
