+++ api/synthgenie/synthesizers/shared/schemas/agent.py
